[DCOS-41589] Extracted foldered Spark tests into a separate module
